don't allow deletion in control points mode

+++ assignments/a3-splines/curve_editor.cpp
@@ -39,6 +39,7 @@ void CurveEditor::addPoint(const vec3& p) {
 }
 
 void CurveEditor::deletePoint(int key) {
+  if (mShowControlPoints) return;
   mSpline.deleteKey(key);
 }
